pyproject.toml
[tool.poetry]
name = "labqoda"
version = "0.0.0"
description = "Labqoda API"
authors = ["Labqoda"]

[tool.poetry.dependencies]
dj-database-url = "^0.5.0"
django = "3.0.9"
django-cors-headers = "^3.3.0"
django-filter = "^2.3.0"
django-localflavor = "^3.0"
django-structlog = "1.5.2"
djangorestframework = "^3.11.0"
djangorestframework-camel-case = "^1.2.0"
drf-spectacular = "^0.9.12"
gunicorn = "^20.0.4"
python = "3.8.*"
ramos = "^1.4.0"
requests = "^2.23.0"
whitenoise = "^5.2.0"
django-storages = "^1.9.1"
django-oauth-toolkit = "^1.3.2"
django-rest-framework-social-oauth2 = "^1.1.0"
social-auth-app-django = "^4.0.0"
django-nested-admin = "^3.3.2"
python-decouple = "^3.3"
boto3 = "^1.14.42"
mysqlclient = "^2.0.1"
Pillow = "^7.2.0"
PyMySQL = "^0.10.0"
django-tinymce = "^3.0.2"
requests-toolbelt = "^0.9.1"
django-taggit = "^1.3.0"
django-rest-passwordreset = "^1.1.0"
drf-writable-nested = "^0.6.1"
django-wkhtmltopdf = "3.3.0"
WeasyPrint = "^51"

[tool.poetry.dev-dependencies]
bandit = "^1.6.2"
black = "^19.10b0"
flake8 = "^3.8.3"
ipython = "^7.11.1"
isort = "^5.2.2"
model_bakery = "^1.0.2"
mypy = "^0.782"
pdbpp = "^0.10.2"
pytest = "^6.0.1"
pytest-cov = "^2.8.1"
pytest-deadfixtures = "^2.1.0"
pytest-django = "^3.9.0"
pytest-env = "^0.6.2"
pytest-pythonpath = "^0.7.3"
pytest-vcr = "^1.0.2"
safety = "^1.8.5"
vcrpy = "^4.1.0"
ipdb = "^0.13.3"

[tool.black]
line-length = 79
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )
)
'''

[tool.isort]
line_length = 79
known_first_party = "labqoda"
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
skip_glob = [
    "**migrations**"
]

[tool.coverage.run]
omit = [
    'src/manage.py',
    '**asgi.py',
    '**wsgi.py',
    '**migrations**',
    '**tests**',
    '**settings**',
]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
